fix typos and small tweaks for readability
There were a few typos I noticed when reading this site online that I
came to fix, and as I scanned the piece again, noticed areas where the
language felt a little clunky and I could improve it with small changes.
So I fixed the typos and the clunks in this commit!