Format chat logs for improved readability
When reading the story on mobile and desktop, I found the chat
logs hard to follow due to the amount of scrolling required.
They are important to the plot, and so I wanted to add a bit of
formatting to make them easier to scan as dialogue.  My aim is
to preserve the feeling of chat logs, while keeping the max-width
of text to the same width as the body text of the piece.

There is still some scrolling required on mobile, but not much.